[GEM] Add guard to get faster feedback on tests
To use it, simply run `bundle exec guard` in a separate terminal. Then
each time the specs or the lib files are updated, the tests are run and a
notification gives you their state (red/green).

I had to move some specs to match the structure of the lib/ directory.