Add pricer and dispatcher as CashRegister dependencies
The goal is to make the Product free of those notions of price and
discount. Those are relevent only to the cash register. We must reflect
that in the code.

The side effect of this is that we don't need a product spec as it
became very simple. Instead we now have a spec for Pricer and one for
Discounter.

We use default values for those CashRegister's arguments as we'll be
able to leave the calls to it in Session unchanged.

The Session spec is doing a better integration test than our previous
integration test thus we better remove it.